Fixed a test, Removed signatures, added gitignore
Fixed a bugged unit test, the error was absorbed by a try catch, so the test
was green but shouldn't have.
Removing useless signatures on test dlls, we have to keep the signature for
test dlls that are set as friendly by the dlls they test.
Removing some useless static functions on WeakTypeFinder.
adding a .gitignore file