Numerous enhancements to the ActivityLogger.
- New ActivityLoggerConsoleSync that displays nice logs. This is
a sample of what could be done with the ActivityLogger (it is
not optimized at all). To be used for tests.
- Full support of exceptions logging: exceptions are now routed
up to the Syncs (inside Groups), and numerous overloads make
it easy to use.
- Group conclusions are now managed explicitly instead of relying
on string concatenations (see ActivityLogGroupConclusion). Since
conclusions are objects, a Sync can handle specfic types it may know
to highlight/process/etc. a conclusion. (It may also just call ToString
to obtain a string for the conclusion).
- Fix stupid error in ErrorCatcher (it was not scoped).