Fixing a test and removing MapType from ISimpleTypeFinder (and its implementations)
The CheckInjection test used to fail silently, because an assert was done in the Teardown function of one of the plugins.
That assert checked that plugins using a service could be warned whenever the service's status changed and that an optional service was stopped when the plugins that reference it were stopped (which is not the case unless you set the stopLaunchedOptionals on the Apply method). To enable this test, it now calls Apply with the correct parameter.